+++ packages/ai/src/embed/index.ts
@@ -1,4 +1,5 @@
 export * from './embed';
+export * from './embed-events';
 export * from './embed-many';
 export * from './embed-many-result';
 export * from './embed-result';