+++ packages/lmnt/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/lmnt
 
+## 3.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [776b617]
+  - @ai-sdk/provider-utils@5.0.0-beta.5
+  - @ai-sdk/provider@4.0.0-beta.3
+
 ## 3.0.0-beta.4
 
 ### Patch Changes