+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/react
 
+## 4.0.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies [ff9ce30]
+  - ai@7.0.0-beta.25
+
+## 4.0.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [776b617]
+  - @ai-sdk/provider-utils@5.0.0-beta.5
+  - ai@7.0.0-beta.24
+
 ## 4.0.0-beta.23
 
 ### Patch Changes