+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [ff9ce30]
+  - ai@7.0.0-beta.25
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [776b617]
+  - ai@7.0.0-beta.24
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [80d4dde]
   - ai@7.0.0-beta.23