+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/valibot
 
+## 3.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [776b617]
+  - @ai-sdk/provider-utils@5.0.0-beta.5
+
 ## 3.0.0-beta.4
 
 ### Patch Changes